Initial Restructure of App into seperate files
Move each component into own component js files in sub directories
under components.
Remove the Weather wrapper class as unneeded.
