+++ src/App.js
@@ -1,255 +1,12 @@
 import React, { Component } from 'react';
+import LocationSearch from './components/LocationSearch/LocationSearch';
+import Forecast from './components/Forecast/Forecast';
+
 //import logo from './logo.svg';
 import './App.css';
 
-class LocationForm extends Component {
+class App extends Component {
 
-  constructor(props) {
-    super(props);
-
-    this.onSearchValueChange = this.onSearchValueChange.bind(this);
-  }
-
-  onSearchValueChange(event) {
-    this.props.searchForLocations(event.target.value);
-  }
-
-  render() {
-    return (
-      <form className="location-form">
-        <div className="form-input">
-          <label htmlFor="location" className="sr-only">Location</label>
-          <input type="text" name="location" className="form-text" placeholder="Search for a city" onChange={this.onSearchValueChange}></input>
-        </div>
-      </form>
-    );
-  }
-}
-
-class LocationList extends Component {
-
-  render() {
-    const resultsMarkup = this.props.searchResults.map((row, index) => {
-      return (
-        <li key={index}>
-          <a href={'#'+row.woeid} onClick={(e) => {
-            e.preventDefault();
-            this.props.changeCity(row.woeid)
-          }} className={index == this.props.chosenResult? 'active' : null}>{row.title}</a>
-        </li>
-      );
-    });
-
-    return (
-      <ul className="location-list">
-        {resultsMarkup}
-      </ul>
-    );
-  }
-}
-
-class LocationSearch extends Component {
-
-  constructor(props) {
-    super(props);
-    this.state = {
-      searchValue: '',
-      searchResults: [],
-      chosenResult: -1
-    };
-
-    this.searchForLocations = this.searchForLocations.bind(this);
-  }
-
-  searchForLocations(query) {
-    fetch('https://api-proxy.newmediathinking.com/metaweather.php?method=location/search&query='+encodeURI(query))
-      .then(response => response.json())
-      .then(data => {
-        this.setState({
-          searchValue: query,
-          searchResults: data.slice(0, 10),
-          chosenResult: -1
-        });
-      });
-  }
-
-  handleKeyDownArrow() {
-    const chosenResult = parseInt(this.state.chosenResult);
-    if (!isNaN(chosenResult) && chosenResult < this.state.searchResults.length - 1) {
-      this.setState({chosenResult: chosenResult + 1});
-    }
-  }
-
-  handleKeyUpArrow() {
-    const chosenResult = parseInt(this.state.chosenResult);
-    if (!isNaN(chosenResult) && chosenResult > -1) {
-      this.setState({chosenResult: chosenResult - 1});
-    }
-  }
-
-  handleEnterKey() {
-    const chosenResult = parseInt(this.state.chosenResult);
-    if (!isNaN(chosenResult) && typeof(this.state.searchResults[chosenResult].woeid) !== 'undefined') {
-      this.props.changeCity(this.state.searchResults[chosenResult].woeid);
-    }
-  }
-
-  render() {
-
-    return (
-      <div className="location-search" onKeyDown={(e) => {
-          if (e.keyCode !== 13 && e.keyCode !== 38 && e.keyCode !== 40) {
-            return;
-          }
-          e.preventDefault();
-          if (e.keyCode === 13){
-            // Enter
-            this.handleEnterKey();
-          } else if (e.keyCode === 38) {
-            // Up Arrow
-            this.handleKeyUpArrow();
-          } else if (e.keyCode === 40) {
-            // Down Arrow
-            this.handleKeyDownArrow();
-          }
-        }}>
-        <LocationForm searchForLocations={this.searchForLocations} ref="locationForm" />
-        <LocationList searchResults={this.state.searchResults} changeCity={this.props.changeCity} chosenResult={this.state.chosenResult} ref="locationList" />
-        <button className="cancel-button" onClick={this.props.changeViewFromClick}>Cancel</button>
-      </div>
-    );
-
-  }
-}
-
-class ForecastDetail extends Component {
-
-  getForecastDate() {
-    let todayDate = new Date();
-    let forecastDate = new Date(this.props.weather.applicable_date);
-
-    return todayDate.setHours(0,0,0,0) === forecastDate.setHours(0,0,0,0) ? 'Today' : forecastDate.toLocaleDateString('en-US', { weekday: 'long' });
-  }
-
-  getWeatherIcon() {
-    const icons = {
-      sn: 'snow',
-      sl: 'snow',
-      h: 'hail',
-      t: 'lightning',
-      hr: 'rain',
-      lr: 'drizzle',
-      s: 'rain',
-      hc: 'fog',
-      lc: 'cloud',
-      c: 'sun'
-    }
-
-    return './icons/weather-'+icons[this.props.weather.weather_state_abbr]+'.svg';
-  }
-
-  getWeatherStateName() {
-    if (this.props.weather.weather_state_name == 'Clear' && parseFloat(this.props.weather.the_temp) >= 15 ) {
-      return 'Sunny';
-    }
-    return this.props.weather.weather_state_name;
-  }
-
-  render() {
-
-    if (typeof(this.props.weather) === 'undefined') {
-      return null;
-    } else {
-      return (
-        <div className="forecast-detail">
-          <h2 className="forecast-day">{this.getForecastDate()}</h2>
-          <img className="forecast-icon" src={this.getWeatherIcon()} />
-          <p className="forecast-text">{this.getWeatherStateName()}</p>
-          <p className="forecast-temp">{parseFloat(this.props.weather.the_temp).toFixed(1)}&deg;C</p>
-        </div>
-      );
-    }
-  }
-}
-
-class ForecastDateSelect extends Component {
-
-  getShortDate(forecastDateStr) {
-
-    const forecastDate = new Date(forecastDateStr);
-    return forecastDate.toLocaleDateString('en-US', { weekday: 'narrow' });
-  }
-
-  render() {
-    if (!Array.isArray(this.props.allWeather)) {
-      return null;
-    }
-
-    const dateSelectMarkup = this.props.allWeather.map((row, index) => {
-      return (
-        <li key={index}><a href={'#day'+index} onClick={(e) => {
-          e.preventDefault();
-          this.props.changeDay(index);
-        }} className={index == this.props.selectedDay? 'active' : null}>{this.getShortDate(row.applicable_date)}</a></li>
-      );
-    });
-
-    return (
-      <ul className="forecast-date-select">
-        {dateSelectMarkup}
-      </ul>
-    );
-  }
-}
-
-class Forecast extends Component {
-
-  constructor(props) {
-    super(props);
-    this.state = {
-      location: '---',
-      weather: [],
-      day: 0
-    };
-
-    this.changeDay = this.changeDay.bind(this);
-  }
-
-  changeDay(newDay) {
-    if (typeof(this.state.weather[newDay]) !== 'undefined') {
-      this.setState({day: newDay});
-    }
-  }
-
-  getWeather() {
-    fetch('https://api-proxy.newmediathinking.com/metaweather.php?method=location/'+this.props.woeid)
-      .then(response => response.json())
-      .then(data => {
-        this.setState({
-            location: data.title,
-            weather: data.consolidated_weather
-          });
-        });
-  }
-
-  componentWillMount() {
-    this.getWeather();
-  }
-
-  render() {
-    const displayWeather = this.state.weather[this.state.day];
-
-    return (
-      <div className="forecast">
-        <h2 className="location-name"><a href="#" onClick={this.props.changeViewFromClick}>{this.state.location}</a></h2>
-        <ForecastDetail weather={displayWeather} />
-        <ForecastDateSelect allWeather={this.state.weather} selectedDay={this.state.day} changeDay={this.changeDay} />
-      </div>
-    );
-  }
-}
-
-class Weather extends Component {
   constructor(props) {
     super(props);
     this.state = {
@@ -294,22 +51,14 @@ class Weather extends Component {
 
   render() {
     return (
-      <div className="weather">
-        {this.renderActiveView()}
-      </div>
-    );
-  }
-}
-
-class App extends Component {
-  render() {
-    return (
       <div className="App">
         <header className="App-header">
           <h1>React Weather</h1>
         </header>
         <div className="App-body">
-          <Weather />
+          <div className="weather">
+            {this.renderActiveView()}
+          </div>
         </div>
         <footer className="App-footer">
           <p>&copy; Andy Broomfield 2018</p>
